refactor: index.tsx 에서 router, sentry, startMocker 분리
