+++ frontend/src/index.tsx
@@ -1,117 +1,35 @@
 import { Global, ThemeProvider } from '@emotion/react';
 import * as Sentry from '@sentry/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import React, { lazy, Suspense } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 
-import App from '@/App';
-
-import { ErrorSuspenseContainer } from './components';
-import { API_ERROR_MESSAGE, ROUTE_PARAM } from './constants';
-import { ROUTE } from './constants/route';
+import router from './router';
 import globalStyles from './styles/globalStyles';
 import theme from './styles/theme';
+import { initializeSentry, retryQuery, startMockWorker } from './utils';
 
-const isProduction = process.env.NODE_ENV === 'production';
-const baseUrlPattern = new RegExp(`^${process.env.API_BASE_URL?.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&')}`);
-
-const HomePage = lazy(() => import('@/pages/HomePage'));
-const DetailedReviewPage = lazy(() => import('@/pages/DetailedReviewPage'));
-const ErrorPage = lazy(() => import('@/pages/ErrorPage'));
-const ReviewListPage = lazy(() => import('@/pages/ReviewListPage'));
-const ReviewWritingCompletePage = lazy(() => import('@/pages/ReviewWritingCompletePage'));
-const ReviewWritingPage = lazy(() => import('@/pages/ReviewWritingPage'));
-const ReviewZonePage = lazy(() => import('@/pages/ReviewZonePage'));
-const ReviewCollectionPage = lazy(() => import('@/pages/ReviewCollectionPage'));
-
-const LoadingPage = lazy(() => import('@/pages/LoadingPage'));
-
-Sentry.init({
-  dsn: `${process.env.SENTRY_DSN}`,
-  enabled: isProduction,
-  integrations: [Sentry.browserTracingIntegration()],
-  environment: 'production',
-  tracesSampleRate: 1.0,
-  tracePropagationTargets: [baseUrlPattern],
-});
-
-export function retryFunction(failureCount: number, error: Error): boolean {
-  const { message } = error;
-  const isServerError = message === API_ERROR_MESSAGE.serverError;
-
-  // Fetch API로 인해 발생한 오류인지 확인
-  // 500번대 에러이면 한 번 더 재시도
-  if (isServerError) return failureCount < 1;
-
-  return false; // 그 외의 경우 재시도하지 않음
-}
+initializeSentry();
 
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
       throwOnError: true,
-      retry: retryFunction,
+      retry: retryQuery,
       refetchOnWindowFocus: false,
     },
     mutations: {
       throwOnError: true,
-      retry: retryFunction,
+      retry: retryQuery,
     },
   },
 });
 
-const router = createBrowserRouter([
-  {
-    path: ROUTE.home,
-    element: (
-      <Suspense fallback={<LoadingPage />}>
-        <App />
-      </Suspense>
-    ),
-    errorElement: <ErrorPage />,
-    children: [
-      {
-        path: '',
-        element: <HomePage />,
-      },
-      { path: `${ROUTE.reviewWriting}/:${ROUTE_PARAM.reviewRequestCode}`, element: <ReviewWritingPage /> },
-      {
-        path: `${ROUTE.reviewWritingComplete}/:${ROUTE_PARAM.reviewRequestCode}`,
-        element: <ReviewWritingCompletePage />,
-      },
-      {
-        path: `${ROUTE.reviewList}/:${ROUTE_PARAM.reviewRequestCode}`,
-        element: <ReviewListPage />,
-      },
-      {
-        path: `${ROUTE.detailedReview}/:${ROUTE_PARAM.reviewRequestCode}/:${ROUTE_PARAM.reviewId}`,
-        element: <DetailedReviewPage />,
-      },
-      {
-        path: `${ROUTE.reviewZone}/:${ROUTE_PARAM.reviewRequestCode}`,
-        element: (
-          <ErrorSuspenseContainer>
-            <ReviewZonePage />
-          </ErrorSuspenseContainer>
-        ),
-      },
-      { path: `${ROUTE.reviewCollection}/:${ROUTE_PARAM.reviewRequestCode}`, element: <ReviewCollectionPage /> },
-    ],
-  },
-]);
-
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 
-async function enableMocking() {
-  if (!isProduction) {
-    const { worker } = await import('./mocks/browser');
-    return worker.start();
-  }
-}
-
-enableMocking().then(() => {
+startMockWorker().then(() => {
   root.render(
     <React.StrictMode>
       <QueryClientProvider client={queryClient}>
