+++ frontend/src/utils/index.ts
@@ -5,7 +5,10 @@ export { default as hasFinalConsonant } from './hasFinalConsonant';
 export { default as substituteString } from './substituteString';
 export { default as calculateParticle } from './calculateParticle';
 export { default as isTouchDevice } from './touchDevice';
+export { default as retryQuery } from './queryRetrier';
+export { default as startMockWorker } from './mockWorkerStarter';
 export * from './date';
 export * from './media';
 export * from './highlight/index';
 export * from './testUtils';
+export { default as initializeSentry } from './analytics/sentry';
