+++ frontend/src/utils/mockWorkerStarter.ts
@@ -1,0 +1,8 @@
+const startMockWorker = async () => {
+  if (process.env.NODE_ENV === 'production') return;
+
+  const { worker } = await import('../mocks/browser');
+  worker.start();
+};
+
+export default startMockWorker;
